Client code in server project *Fixed Server code in Client project *Fixed
